GCI-1428: Reduce Kafka producer code duplication.
