+++ src/main/java/uk/gov/companieshouse/itemhandler/kafka/OrdersKafkaProducer.java
@@ -1,23 +1,15 @@
 package uk.gov.companieshouse.itemhandler.kafka;
 
-import java.util.Map;
-import java.util.concurrent.ExecutionException;
-import org.springframework.beans.factory.InitializingBean;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import uk.gov.companieshouse.itemhandler.logging.LoggingUtils;
-import uk.gov.companieshouse.kafka.exceptions.ProducerConfigException;
 import uk.gov.companieshouse.kafka.message.Message;
-import uk.gov.companieshouse.kafka.producer.Acks;
-import uk.gov.companieshouse.kafka.producer.CHKafkaProducer;
 import uk.gov.companieshouse.kafka.producer.ProducerConfig;
 
-@Service
-public class OrdersKafkaProducer implements InitializingBean {
+import java.util.Map;
+import java.util.concurrent.ExecutionException;
 
-    private CHKafkaProducer chKafkaProducer;
-    @Value("${spring.kafka.consumer.bootstrap-servers}")
-    private String brokerAddresses;
+@Service
+public class OrdersKafkaProducer extends KafkaProducer {
 
     /**
      * Sends message to Kafka topic
@@ -28,25 +20,11 @@ public class OrdersKafkaProducer implements InitializingBean {
     public void sendMessage(final Message message) throws ExecutionException, InterruptedException {
         Map<String, Object> logMap = LoggingUtils.createLogMapWithKafkaMessage(message);
         LoggingUtils.getLogger().info("Sending message to kafka topic", logMap);
-        chKafkaProducer.send(message);
+        getChKafkaProducer().send(message);
     }
 
     @Override
-    public void afterPropertiesSet() throws Exception {
-        LoggingUtils.getLogger().trace("Configuring CH Kafka producer");
-        ProducerConfig config = new ProducerConfig();
-        if (brokerAddresses != null && !brokerAddresses.isEmpty()) {
-            config.setBrokerAddresses(brokerAddresses.split(","));
-        } else {
-            throw new ProducerConfigException("Broker addresses for kafka broker missing, check if environment variable KAFKA_BROKER_ADDR is configured. " +
-                    "[Hint: The property 'kafka.broker.addresses' uses the value of this environment variable in live environments " +
-                    "and that of 'spring.embedded.kafka.brokers' property in test.]");
-        }
-
-        config.setRequestTimeoutMilliseconds(3000);
-        config.setRoundRobinPartitioner(true);
-        config.setAcks(Acks.WAIT_FOR_ALL);
-        config.setRetries(10);
-        chKafkaProducer = new CHKafkaProducer(config);
+    protected void modifyProducerConfig(final ProducerConfig producerConfig) {
+        producerConfig.setRequestTimeoutMilliseconds(3000);
     }
 }
